Add solution for day 9 part 2
Refactor the `Rope` struct to allow a const generic parameter `N`
specifying the number of knots.  Note that the key is the four added
arms to the match expression when adjusting the next knot.  This is
because the head only moves in four directions parallel to each of the
axes in both directions.  But subsequent knots can also move diagonally.
This adds more more cases to consider for the difference between
successive knots after the leader has moved.