Add solution for day 19 part 2
This solution is *not* guaranteed to find an optimal solution.  I use a
breadth-first search with a fixed limit (here 1000) on the number of
partial solutions to keep at each generation.  Partial solutions are
sorted by the number of robots, then resources, of each kind in the
order geode, obsidian, clay, ore.

Keeping just the top 1000 solutions at each step seems to be more than
adequate to solve both the example and input.