Day 18: Refactor solution to make more use of `ndarray`
This version uses more features of the `ndarray` crate in an effort to
reduce code duplication.  By using the views, slices, and iterators it
provides, we can largely eliminate the use of indices to access array
elements which often leads to highly repetitive code.

This version also refactors the function which counts boundary faces to
use arrays instead of hash maps.  We still use a hash set as a way to
dedupe the list of cubes, as the counting method will fail if the list
contains the a cube more than once.