Day 15: Fix bug in part 2 and refactor code
It looks like there was a bug in the solution to part 2 in the case
where the union of the disks from each sensor report forms a single
closed interval on a given row.  It is ok if that interval's start
position is greater than the minimum of the x-range, or if it's end
position is less than the maximum of the x-range, as long as either of
those is by at most 1 and not both of them occur.  So the check seems
easier to express by summing the lengths of the gaps between x-range
endpoint and the max/min of the interval's and x-range's endpoints.

Fortunately, this bug didn't affect the solution to either the example
of the puzzle input.

Also refactor the code somewhat.  Add a method to `SensorReport` to take
a row and return the intersection of the report's l1-disk, centered at
the sensor and having radius equal to the distance from that sensor to
the nearest beacon, with that row.  This eliminates the need for what
was the fn `build_intervals()`.

In `main.rs`, we need to provide some additional parameters for solving
each of the parts of the puzzle: the row in part 1 and the x and y
ranges to search for a unique beacon position in part 2.  These
parameters are different for the examples provided in the puzzle
description and the real puzzle input.  Rather than requiring additional
command-line arguments, let's keep it simple by just looking at the
parsed input and determining from that whether we were given the example
of the real input and use the appropriate parameters.  Not only does
this make the API consistent with other days, but it eliminates the need
for a user to have those additional parameters handy.