Day 17: Improve solution to part 2 by using "effective board"
My original solution to part 2 worked to solve the given puzzle input,
but it did not get a solution for the provided example.  The reason is
that the example did not result in any future rows all of whose cells
were filled with rock, except for the original floor.  So it was asking
too much to find such a row, although such a row did recur in the
periodic behavior of the puzzle input.

In this version of the solution, we make use of the notion of the
"effective board".  This is one in which open cells which cannot be
reached by any future falling rocks might as well be filled.  By filling
in these cells, we do find additional floors (i.e., rows all of whose
cells are filled) in the periodic behavior of the example.