Day 16: Improve speed of solutions
Both parts of this problem benefit from using a suboptimal beam search
algorithm to quickly find a good solution and then use this to help
early pruning during an optimal backtrack algorithm.

The beam search can find the optimal solution to both the example and
the puzzle input with a stack size of 1000, although we don't know for
sure it is optimal.  Increasing the stack size about 1000 does not
change the answer it finds, so it is likely the best.  But we can
combine this with a complete backtracking algorithm, using the score of
the solution found by beam search to abandon partial solutions early.

Also use `rayon` to explore the top of the search tree in the
backtracking algorithm concurrently with multiple threads.