Day 19: Find optimal solution with exhaustive tree search
This solution is a breadth-first tree search with pruning of partial
solutions that cannot be optimal.  It runs in a reasonable amount of
time without using the beam search limit on the number of partial
solutions to retain at each generation.  This is made possible by
several insights allowing more pruning.

After sorting partial solutions using the heuristic in the impl of `Ord`
for `State`, we can dedup the vector, where we consider a solution a
"duplicate" if it is dominated by another.  This is fast after paying
for sorting and helps reduce the length of the next level of partial
solutions considerably.

Another insight allows us limit when we extend partial solutions by not
purchasing any robot in a given minute.  We should only do this when
there is some robot which we currently cannot afford, but which we will
eventually be able to afford if we wait (meaning that we have the
necessary robots to produce the resources required to purchase that
other robot).  If we will not be able to eventually afford a robot which
we cannot already afford given what robots we own, then it would not
make sense to not purchase one of the robots we can afford now.  Waiting
and then buying later only means it will make less resources over the
remaining time.  So an optimal solution would never do this.

Finally, it is useless to own more robots of a given type than the
maximum value of the corresponding resource necessary to purchase
another robot type.  This is because in a single minute, we can only
purchase one robot.  Thus, if we had more of one type of robot than we
can spend in a minute, we are producing more of that resource than we
will ever be able to spend.  Therefore, excluding geode-robots, the
other types of robots have a limit and we should not consider solutions
that eventually own more than this number of that robot kind.

Using these insights to reduce how many extensions of each partial
solution we consider, the sizes of the generations remains manageable
even out to depth 32.