Day 19: Further improve performance
This version precomputes the limit on the number of each kind of robot
an optimal solution can own, instead of computing that limit each time
we try to advance a partial solution.  That nearly doubles the speed.

It also uses a suboptimal, fast beam search to first find a good
solution and then use that to further prune the exhaustive search tree.
We add a very crude but fast upper bound on the number of geodes that a
`State` could ever produce in a given amount of time.  Even though this
is often a vast overestimate, it suffices to help substantially prune
the lower levels of the tree.