Day 19: Return `usize` instead of `Option<State>`
There is no reason to return an option; if the time limit is 0 then the
default `State` is the right thing to return anyway.

Also, we cannot use a `State` to reconstruct the sequence of purchasing
decisions.  So there is really no reason to return `State`.  Either we
keep track of history (which will slow things down and require lots of
allocations) so as to be able to build the optimal strategy, or we just
return the optimal number of geodes.  Here we choose the later.