Day 22: Add part2 solution to main program
I had not previously added the solution to part 2 to the main program,
instead just using the tests in the library to obtain and submit the
answer.  We have hardcoded edge gluing rules for the example and actual
inputs, so in order to solve part 2 in `main.rs` we have to guess which
input we were given.  Here we use the shape of the grid to determine
which gluing rules to use.