Day 15 part 2: Fix bugs and use parallelism
The original solution to part 2 was not careful to verify that there was
a unique solution.  It used an early return as soon as it found a line
with a single uncovered point, rather than verifying that there was only
one such line.  It also incorrectly handled the subcases when the union
of neighborhoods intersecting a given line could be written as a single
interval.  This version corrects these bugs.

Using `rayon` to search the `yrange` in parallel is also a substantial
performance improvement.