Day 16: Restrict to subgraph
Refactor the solution to day 16 to limit the walks to the subgraph of
nodes whose valves have positive flow rates.  For my input, this reduces
the number of vertices from 59 to 15.  We precompute the shortest paths
between all pairs of vertices and use those distances when stepping from
one vertex in the subgraph to another.  This vastly speeds up the
search, eliminating many paths that take suboptimal routes between pairs
of vertices with positive flow rates.

The overall strategy is the same as before, namely, use a fast,
suboptimal beam search to find a good solution and then a backtrack with
pruning to find the optimal solution.  The details are quite different
in this approach, however, as we have to account for the different
travel times between various pairs of nodes.

This version is fast enough that multithreading is no longer used.