Day 20: Improve performance
It turns out to be significantly faster to get rid of the permutations
`perm` and `inv` and instead just label each element of the original
slice with its position, then on step `t` of the mixing just search
the for the element that was originally in position `t`.

This version runs in about 52ms compared to about 125ms for the previous
version.