Day 23: Improve performance with 2-d array
On each round, create a 2-d array of `bool`s to indicate the positions
of the elves.  Then use this array when deciding where each elf should
move.  This avoids the many `HashSet::contains()` calls.  Instead, we
iterate over the set (which does not require hashing) and set the
positions of the elves.  After that, lookups are into the array instead
of the `HashSet`.