Day 23: Use `Vec` instead of `HashSet`
All we need is a list of positions where the elves are found, we do not
need to keep them in a set anymore since we no longer use membership
tests.  This improved the speed by about 60ms.

We also store a vector of proposed moves rather than a vector of tuples
`(current, proposed)`, and later we zip the vector of current positions
with the vector of proposed positions.  This improved things by another
3 or 4 ms.

Note that I kept the `HashSet` as the argument to `Elves::new()`, since
we want to make sure the list of positions we are given are unique.  We
could also try taking a vector and sorting and deduping it.