Day 20: Speed up searching for next pivot
Instead of always starting at the beginning of the slice `perm` and
searching forward until we find `t`, start looking for `t` where we last
moved it on the previous round.  Then search both forward and backward
from there, interleaved.  This is much faster.

However, the first two rounds are special.  On the first round, the
permutation starts sorted in increasing order.  Two adjacent elements (t
- 1, t) can only be separated when another element u < t - 1 is moved to
position t - 1, shifting t - 1 down 1.  Thus, when we are looking for
`t`, we can start looking for it wherever we found `t - 1` and only
search forward from there.  So in round 0 use that variation.

Round 1 is also special because of the quirks of round 0.  In round 0,
initially many of the early elements are found at index 0.  This means
that anything shifted by moving these first elements is shifted down.
So in round 1, we tend to find low index elements at positions much
lower than where we left them in round 0.  Eventually that wears off and
later we find the highest indices at positions much greater than where
they were left in round 0.  Furthermore, it tends to be that the offset
between where we left an element in round 0 and where we find it in
round 1 is similar for elements whose indices differ by little.  So in
round 1, our strategy is to guess how much to offset `last[t]` when
starting to look for `t` based on what we found on the previous
iteration for `t - 1`.

Finally, remove the test `test_mix()`, as it is getting cumbersome to
keep updating it every time I refactor the signature of `mix()`.