Day 15: Replace area search with perimeter search
Do not solve part 2 by redoing part 1 for each horizontal line segment
in the search region, effectively searching the entire area.  It is a
*significant* improvement to only search along the perimeters of the
discs for points not covered by any disc.

If there are any points in the search area not covered by any of the
sensor report discs, then there is a point not covered by any disc which
is immidately adjacent to a disc.  So it is enough to search each line
segment along the "exterior perimeter" of each of the discs.  So instead
of solving part 1 4M times, we do it 4 * 19 = 76 times.  The slight
adjustment from part 1 is that we are not intersecting discs with
horizontal lines but rather with lines of slope +/- 1.